feat(apig): resources support enable_force_new param

+++ huaweicloud/services/apig/resource_huaweicloud_apig_application_authorization.go
@@ -10,6 +10,7 @@ import (
 	"github.com/hashicorp/go-multierror"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 
 	"github.com/chnsz/golangsdk"
 
@@ -19,6 +20,12 @@ import (
 )
 
 var strSliceParamKeys = []string{"api_ids"}
+
+var applicationAuthorizationNonUpdatableParams = []string{
+	"instance_id",
+	"application_id",
+	"env_id",
+}
 
 // @API APIG DELETE /v2/{project_id}/apigw/instances/{instance_id}/app-auths/{app_auth_id}
 // @API APIG POST /v2/{project_id}/apigw/instances/{instance_id}/app-auths
@@ -34,6 +41,8 @@ func ResourceApplicationAuthorization() *schema.Resource {
 			StateContext: resourceApplicationAuthorizationImportState,
 		},
 
+		CustomizeDiff: config.FlexibleForceNew(applicationAuthorizationNonUpdatableParams),
+
 		Schema: map[string]*schema.Schema{
 			"region": {
 				Type:        schema.TypeString,
@@ -45,19 +54,16 @@ func ResourceApplicationAuthorization() *schema.Resource {
 			"instance_id": {
 				Type:        schema.TypeString,
 				Required:    true,
-				ForceNew:    true,
 				Description: "The ID of the dedicated instance to which the application and APIs belong.",
 			},
 			"application_id": {
 				Type:        schema.TypeString,
 				Required:    true,
-				ForceNew:    true,
 				Description: "The ID of the application authorized to access the APIs.",
 			},
 			"env_id": {
 				Type:        schema.TypeString,
 				Required:    true,
-				ForceNew:    true,
 				Description: "The environment ID where the APIs were published.",
 			},
 			"api_ids": {
@@ -80,6 +86,18 @@ the new value next time the change is made. The corresponding parameter name is 
 						Internal: true,
 					},
 				),
+			},
+
+			// Internal parameters.
+			"enable_force_new": {
+				Type:         schema.TypeString,
+				Optional:     true,
+				ValidateFunc: validation.StringInSlice([]string{"true", "false"}, false),
+				Description: utils.SchemaDesc(
+					`Whether to allow parameters that do not support changes to have their change-triggered behavior set to 'ForceNew'.`,
+					utils.SchemaDescInput{Internal: true,
+						Required: true,
+					}),
 			},
 		},
 	}
@@ -267,6 +285,7 @@ func resourceApplicationAuthorizationRead(_ context.Context, d *schema.ResourceD
 	return nil
 }
 
+// nolint:gocyclo
 func resourceApplicationAuthorizationUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	cfg := meta.(*config.Config)
 	client, err := cfg.ApigV2Client(cfg.GetRegion(d))
@@ -274,50 +293,52 @@ func resourceApplicationAuthorizationUpdate(ctx context.Context, d *schema.Resou
 		return diag.Errorf("error creating APIG v2 client: %s", err)
 	}
 
-	var (
-		resourceId = d.Id()
-		instanceId = d.Get("instance_id").(string)
-		appId      = d.Get("application_id").(string)
-		envId      = d.Get("env_id").(string)
+	if d.HasChangeExcept("enable_force_new") {
+		var (
+			resourceId = d.Id()
+			instanceId = d.Get("instance_id").(string)
+			appId      = d.Get("application_id").(string)
+			envId      = d.Get("env_id").(string)
 
-		consoleApiIds, scriptApiIds = d.GetChange("api_ids")
+			consoleApiIds, scriptApiIds = d.GetChange("api_ids")
 
-		consoleApiIdsList = consoleApiIds.(*schema.Set).List()
-		scriptApiIdsList  = scriptApiIds.(*schema.Set).List()
-		originApiIdsList  = d.Get("api_ids_origin").([]interface{})
-	)
+			consoleApiIdsList = consoleApiIds.(*schema.Set).List()
+			scriptApiIdsList  = scriptApiIds.(*schema.Set).List()
+			originApiIdsList  = d.Get("api_ids_origin").([]interface{})
+		)
 
-	// Lock the resource to prevent concurrent updates (error APIG.3500 will be returned if the etcd data synchronize
-	// failed)
-	config.MutexKV.Lock(resourceId)
-	defer config.MutexKV.Unlock(resourceId)
+		// Lock the resource to prevent concurrent updates (error APIG.3500 will be returned if the etcd data synchronize
+		// failed)
+		config.MutexKV.Lock(resourceId)
+		defer config.MutexKV.Unlock(resourceId)
 
-	newApiIds := utils.FindSliceElementsNotInAnother(scriptApiIdsList, consoleApiIdsList)
-	rmApiIds := utils.FindSliceElementsNotInAnother(originApiIdsList, scriptApiIdsList)
+		newApiIds := utils.FindSliceElementsNotInAnother(scriptApiIdsList, consoleApiIdsList)
+		rmApiIds := utils.FindSliceElementsNotInAnother(originApiIdsList, scriptApiIdsList)
 
-	if len(rmApiIds) > 0 {
-		log.Printf("[DEBUG] Prepare to delete the authorization for specified API IDs: %v", rmApiIds)
-		err := deleteApplicationAuthorizationForApis(client, instanceId, envId, appId, rmApiIds)
-		if err != nil {
-			return diag.FromErr(err)
+		if len(rmApiIds) > 0 {
+			log.Printf("[DEBUG] Prepare to delete the authorization for specified API IDs: %v", rmApiIds)
+			err := deleteApplicationAuthorizationForApis(client, instanceId, envId, appId, rmApiIds)
+			if err != nil {
+				return diag.FromErr(err)
+			}
 		}
-	}
 
-	if len(newApiIds) > 0 {
-		log.Printf("[DEBUG] Prepare to create the authorization for specified API IDs: %v", newApiIds)
-		err = createApplicationAuthorizationForApis(client, instanceId, envId, appId, newApiIds)
-		if err != nil {
-			return diag.FromErr(err)
+		if len(newApiIds) > 0 {
+			log.Printf("[DEBUG] Prepare to create the authorization for specified API IDs: %v", newApiIds)
+			err = createApplicationAuthorizationForApis(client, instanceId, envId, appId, newApiIds)
+			if err != nil {
+				return diag.FromErr(err)
+			}
 		}
-	}
 
-	// If the request is successful, obtain the values of all slice parameters first and save them to the corresponding
-	// '_origin' attributes for subsequent determination and construction of the request body during next updates.
-	// And whether corresponding parameters are changed, the origin values must be refreshed.
-	err = utils.RefreshSliceParamOriginValues(d, strSliceParamKeys)
-	if err != nil {
-		// Don't fail the update if origin refresh fails
-		log.Printf("[WARN] Unable to refresh the origin values: %s", err)
+		// If the request is successful, obtain the values of all slice parameters first and save them to the corresponding
+		// '_origin' attributes for subsequent determination and construction of the request body during next updates.
+		// And whether corresponding parameters are changed, the origin values must be refreshed.
+		err = utils.RefreshSliceParamOriginValues(d, strSliceParamKeys)
+		if err != nil {
+			// Don't fail the update if origin refresh fails
+			log.Printf("[WARN] Unable to refresh the origin values: %s", err)
+		}
 	}
 
 	return resourceApplicationAuthorizationRead(ctx, d, meta)
